+++ spec/systems/ConditionalVisibilityPf2e.spec.ts
@@ -25,7 +25,7 @@ describe('ConditionalVisibilitySystem5e', () => {
 
     describe('Setup', () => {
         it('Establishes one condition for pf2e', () => {
-            const effects:Map<String, String> = system.effects();
+            const effects:Map<string, string> = system.effectsByIcon();
             expect(effects.size).toBe(1);
             expect(effects.get('systems/pf2e/icons/conditions-2/invisible.png')).toBe('invisible');
         });