+++ src/module/systems/DefaultConditionalVisibilitySystem.ts
@@ -1,3 +1,4 @@
+import { ConditionalVisibilityFacade } from '../ConditionalVisibilityFacade';
 import * as Constants from '../Constants';
 import { ConditionalVisibilitySystem } from "./ConditionalVisibilitySystem";
 
@@ -6,12 +7,16 @@ import { ConditionalVisibilitySystem } from "./ConditionalVisibilitySystem";
  */
 export class DefaultConditionalVisibilitySystem implements ConditionalVisibilitySystem {
     
-    _effects:Map<String, String>;
+    _effectsByIcon: Map<string, string>;
+    _effectsByCondition: Map<string, string>;
 
     constructor() {
-        this._effects = new Map<String, String> ([['modules/conditional-visibility/icons/unknown.svg', 'invisible'],
-        ['modules/conditional-visibility/icons/foggy.svg', 'obscured'],
-        ['modules/conditional-visibility/icons/moon.svg', 'indarkness']]);
+        //yes, this is a BiMap but the solid TS BiMap implementaiton is GPLv3, so we will just fake what we need here
+        this._effectsByIcon = this.effects();
+        this._effectsByCondition = new Map();
+        this._effectsByIcon.forEach((value: string, key: string) => {
+            this._effectsByCondition.set(value, key);
+        });
     }
 
     gameSystemId(): string {
@@ -21,16 +26,35 @@ export class DefaultConditionalVisibilitySystem implements ConditionalVisibility
     /**
      * Base effects are invisible, obscured, and indarkness
      */
-    public effects():Map<String, String> {
-        return this._effects;
+    protected effects():Map<string, string> {
+        return new Map<string, string> ([['modules/conditional-visibility/icons/unknown.svg', 'invisible'],
+        ['modules/conditional-visibility/icons/foggy.svg', 'obscured'],
+        ['modules/conditional-visibility/icons/moon.svg', 'indarkness']]);
+    }
+
+
+    public effectsByIcon(): Map<string, string> {
+        return this._effectsByIcon;
+    }
+
+    public effectsByCondition(): Map<string, string> {
+        return this._effectsByCondition;
     }
 
     public initializeStatusEffects():void {
         console.log(Constants.MODULE_NAME + " | Initializing visibility system effects " + this.gameSystemId() + " for game system " + game.system.id);
-        for (const effect of this.effects().keys()) {
+        for (const effect of this.effectsByIcon().keys()) {
             //@ts-ignore
             CONFIG.statusEffects.push(effect);	
         }
+    }
+
+    /**
+     * For subclasses to set up systsem specific hooks.
+     * @todo unify initializeOnToggleEffect if possible
+     */
+    public initializeHooks(facade:ConditionalVisibilityFacade): void {
+
     }
 
     /**
@@ -43,7 +67,7 @@ export class DefaultConditionalVisibilitySystem implements ConditionalVisibility
 
     public recalculateVisibleStatus(token: any, update: any) {
         if (update.effects) {
-            this._effects.forEach((value, key, map) => {
+            this._effectsByIcon.forEach((value, key, map) => {
                 if (!update.flags) {
                     update.flags = {};
                 }
@@ -173,12 +197,16 @@ export class DefaultConditionalVisibilitySystem implements ConditionalVisibility
         return true;
     }
 
+    public hasStealth():boolean {
+        return false;
+    }
+
     /**
      * Roll for the contested hiding check; override in subclass systems
      * @param token the token whose stats may create the roll.
      * @return a Roll
      */
-    protected rollStealth(token:Token):Roll {
+    public rollStealth(token:Token):Roll {
         return new Roll("1d20");
     }
 
@@ -208,7 +236,7 @@ export class DefaultConditionalVisibilitySystem implements ConditionalVisibility
         const content = await renderTemplate("modules/conditional-visibility/templates/stealth_hud.html", { initialValue: result });
         return new Promise((resolve, reject) => {   
             let hud = new Dialog({
-                title: game.i18n.format('CONVIS.hidden', {}),
+                title: game.i18n.localize('CONVIS.hidden'),
                 content: content,
                 buttons: {
                     one: {